requirements.txt
Flask>=3.0
opencv-python>=4.0
Pillow>=10.0
numpy>=1.20
torch>=2.0
torchvision>=0.15
tqdm>=4.0
